// src/model/document.ts
export type Point = { x: number; y: number };
export type Size = { width: number; height: number };